[M]测试内部装潢的资源路径。

+++ saint-coinach/src/main/java/ffxiv/housim/saintcoinach/db/xiv/entity/housing/YardCatalogCategory.java
@@ -14,12 +14,12 @@ public class YardCatalogCategory extends XivRow {
         return asString("Category");// Name
     }
 
-    public Short getHousingItemCategory() {
+    public short getHousingItemCategory() {
         return asInt16("HousingItemCategory");// 17-庭具
     }
 
-    public byte getOrder() {
-        return asByte("Order");
+    public short getOrder() {
+        return asInt16("Order");
     }
 
     @Override
